+++ src/IEnumerable.php
@@ -1,0 +1,10 @@
+<?php declare(strict_types=1);
+
+namespace MF\Collection;
+
+interface IEnumerable extends \IteratorAggregate, \Countable
+{
+    public function count(): int;
+
+    public function getIterator(): iterable;
+}